+++ CHANGELOG.md
@@ -4,6 +4,15 @@
 
 严格遵循 [`Semantic Versioning 2.0.0`](http://semver.org/lang/zh-CN/) 语义化版本规范。
 
+## 5.2.12 未发
+
+- @uiw-admin/components
+  - 🐞`ProTable`: 修复分页错误
+
+## 5.2.11
+
+- @uiw-admin/utils
+  - 🌟 `request`新增一个`requestType`参数
 
 ## 5.2.10